Extract Auth flow helpers
Move flow helper methods into Auth+FlowHelpers.swift.

Set the integration test environment flag in the runner.
